// frontend/generated-source/MyProjectModels.ts
export class SpeakerResponseDto implements ISpeakerResponseDto {
    id?: number;
    fullName!: string | undefined;
    bio!: string | undefined;
    company?: string | undefined;
    jobTitle?: string | undefined;
    profileImageUrl?: string | undefined;
    twitterHandle?: string | undefined;
    linkedInUrl?: string | undefined;
    websiteUrl?: string | undefined;
    totalTalks?: number;

    constructor(data?: ISpeakerResponseDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.fullName = _data["fullName"];
            this.bio = _data["bio"];
            this.company = _data["company"];
            this.jobTitle = _data["jobTitle"];
            this.profileImageUrl = _data["profileImageUrl"];
            this.twitterHandle = _data["twitterHandle"];
            this.linkedInUrl = _data["linkedInUrl"];
            this.websiteUrl = _data["websiteUrl"];
            this.totalTalks = _data["totalTalks"];
        }
    }

    static fromJS(data: any): SpeakerResponseDto {
        data = typeof data === 'object' ? data : {};
        let result = new SpeakerResponseDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["fullName"] = this.fullName;
        data["bio"] = this.bio;
        data["company"] = this.company;
        data["jobTitle"] = this.jobTitle;
        data["profileImageUrl"] = this.profileImageUrl;
        data["twitterHandle"] = this.twitterHandle;
        data["linkedInUrl"] = this.linkedInUrl;
        data["websiteUrl"] = this.websiteUrl;
        data["totalTalks"] = this.totalTalks;
        return data;
    }
}

export interface ISpeakerResponseDto {
    id?: number;
    fullName: string | undefined;
    bio: string | undefined;
    company?: string | undefined;
    jobTitle?: string | undefined;
    profileImageUrl?: string | undefined;
    twitterHandle?: string | undefined;
    linkedInUrl?: string | undefined;
    websiteUrl?: string | undefined;
    totalTalks?: number;
}